requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
Automat==20.2.0
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==3.1
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.0.9
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap-pagination==1.7.1
django-cities-light==3.6.0
django-ckeditor==5.9.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-embed-video==1.3.2
django-filter==2.3.0
django-intl-tel-input==0.3.1
django-js-asset==1.2.2
django-paypal==1.0.0
django-request==1.5.6
django-simple-history==2.11.0
django-smart-selects==1.5.6
django-social-share==1.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
geoip2==4.0.2
gunicorn==20.0.4
httplib2==0.14.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipdetailscache==0.4.8
ipinfo==3.0.0
ipinfo-django==1.1.0
IPy==1.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
maxminddb==2.0.2
monotonic==1.5
multidict==4.7.6
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
panda==0.3.1
pandas==1.1.1
paramiko==2.6.0
parsel==1.6.0
pbr==5.5.0
pexpect==4.6.0
Pillow==7.2.0
progressbar2==3.52.1
Protego==0.1.16
protobuf==3.6.1
proxyscrape==0.3.0
psycopg2==2.8.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
PyDispatcher==2.0.5
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pypi==2.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-utils==2.4.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.5.0
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
Scrapy==2.3.0
scrapy-proxy-pool==0.1.9
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.48.0
systemd-python==234
testresources==2.0.1
toml==0.10.1
Twisted==20.3.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.28
w3lib==1.22.0
wadllib==1.3.3
whitenoise==5.2.0
xkit==0.0.0
yarl==1.5.1
zope.interface==5.1.0
